+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
-### [0.7.6] - 2026-07-25
+### [0.8.0] - 2026-07-26
 
+- replace native resolver with dns2
 - add stub classes server expects: publisher, signer, source, transport
 
 ### [0.7.5] - 2026-04-20
@@ -104,4 +105,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [0.6.0]: https://github.com/NicTool/dns-nameserver/releases/tag/v0.6.0
 [0.7.4]: https://github.com/NicTool/dns-nameserver/releases/tag/v0.7.4
 [0.7.5]: https://github.com/NicTool/dns-nameserver/releases/tag/v0.7.5
-[0.7.6]: https://github.com/NicTool/dns-nameserver/releases/tag/v0.7.6
+[0.8.0]: https://github.com/NicTool/dns-nameserver/releases/tag/v0.8.0